refactor: separate CommitList component
